+++ satella/coding/concurrent/__init__.py
@@ -5,7 +5,7 @@ from .futures import Future, WrappingFuture, InvalidStateError, FutureCollection
 from .id_allocator import IDAllocator, SequentialIssuer
 from .locked_dataset import LockedDataset
 from .locked_structure import LockedStructure
-from .monitor import MonitorList, Monitor, MonitorDict, RMonitor
+from .monitor import MonitorList, Monitor, MonitorDict, RMonitor, MonitorSet
 from .sync import sync_threadpool
 from .thread import TerminableThread, Condition, SingleStartThread, call_in_separate_thread, \
     BogusTerminableThread, IntervalTerminableThread
@@ -17,7 +17,7 @@ __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableT
            'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
            'CallNoOftenThan', 'SingleStartThread', 'IDAllocator', 'call_in_separate_thread',
            'BogusTerminableThread', 'Timer', 'parallel_execute', 'run_as_future',
-           'sync_threadpool', 'IntervalTerminableThread', 'Future',
+           'sync_threadpool', 'IntervalTerminableThread', 'Future', 'MonitorSet',
            'WrappingFuture', 'InvalidStateError', 'PeekableQueue',
            'CancellableCallback', 'ThreadCollection', 'FutureCollection',
            'SequentialIssuer']