+++ satella/random.py
@@ -1,8 +1,29 @@
 import os
 import random
 import typing as tp
+import string
 
-__all__ = ['shuffle_together', 'random_binary']
+__all__ = ['shuffle_together', 'random_binary', 'random_word']
+
+from satella.coding.typing import T
+
+
+def random_word(length: int, choice: T = string.ascii_lowercase,
+                join_fun: tp.Callable[[T, ...], T] = lambda *args: ''.join(args)):
+    """
+    Build and return a random word of provided length.
+
+    The word will be built by calling join_fun with length of arguments picked
+    at random from choice.
+
+    :param length: length of the word
+    :param choice: a range of characters to use. By default is string.ascii_lowercase
+    :param join_fun: an argument to be called with randomly picked characters as *args.
+        Defaults to ''.join(args), so your T must be a string. If you're passing a
+        different type, remember to alter this function because the default one expects strings!
+    :return: a random word
+    """
+    return join_fun(*(random.choice(choice) for _ in range(length)))
 
 
 def random_binary(length: int) -> bytes: