+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.16
+
+* deprecated `call_and_return_stdout`