+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * deprecated `call_and_return_stdout`
 * deprecated `retry(call_on_failure=)`
 * added default parameter to `ThreadCollection`
+* axed entire CallHierarchy