+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * deprecated `retry(call_on_failure=)`
 * added default parameter to `ThreadCollection`
 * axed entire CallHierarchy
+* added __add__ and __iadd__ to `ThreadCollection`