+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.17
+
+* added automatic enum parsing to satella's JSONEncoder