+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added automatic enum parsing to satella's JSONEncoder
 * Satella's JSON file handling will internally use Satella's JSONEncoder
+* added ExponentialBackoff.launch