+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.20
+
+* added is_subset