+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.17.21
-
-* added random_word
+# v2.17.22